+++ controllers/AdminController.php
@@ -11,11 +11,13 @@
 
 namespace dektrium\user\controllers;
 
+use dektrium\user\filters\AccessRule;
 use dektrium\user\Finder;
 use dektrium\user\models\Profile;
 use dektrium\user\models\User;
 use dektrium\user\models\UserSearch;
 use dektrium\user\Module;
+use dektrium\user\traits\EventTrait;
 use Yii;
 use yii\base\ExitException;
 use yii\base\Model;
@@ -37,6 +39,92 @@ use yii\widgets\ActiveForm;
  */
 class AdminController extends Controller
 {
+    use EventTrait;
+
+    /**
+     * Event is triggered before creating new user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_CREATE = 'beforeCreate';
+
+    /**
+     * Event is triggered after creating new user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_CREATE = 'afterCreate';
+
+    /**
+     * Event is triggered before updating existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_UPDATE = 'beforeUpdate';
+
+    /**
+     * Event is triggered after updating existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_UPDATE = 'afterUpdate';
+
+    /**
+     * Event is triggered before updating existing user's profile.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_PROFILE_UPDATE = 'beforeProfileUpdate';
+
+    /**
+     * Event is triggered after updating existing user's profile.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_PROFILE_UPDATE = 'afterProfileUpdate';
+
+    /**
+     * Event is triggered before confirming existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_CONFIRM = 'beforeConfirm';
+
+    /**
+     * Event is triggered after confirming existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_CONFIRM = 'afterConfirm';
+
+    /**
+     * Event is triggered before deleting existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_DELETE = 'beforeDelete';
+
+    /**
+     * Event is triggered after deleting existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_DELETE = 'afterDelete';
+
+    /**
+     * Event is triggered before blocking existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_BLOCK = 'beforeBlock';
+
+    /**
+     * Event is triggered after blocking existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_BLOCK = 'afterBlock';
+
+    /**
+     * Event is triggered before unblocking existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_UNBLOCK = 'beforeUnblock';
+
+    /**
+     * Event is triggered after unblocking existing user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_UNBLOCK = 'afterUnblock';
+
     /** @var Finder */
     protected $finder;
 
@@ -66,13 +154,13 @@ class AdminController extends Controller
             ],
             'access' => [
                 'class' => AccessControl::className(),
+                'ruleConfig' => [
+                    'class' => AccessRule::className(),
+                ],
                 'rules' => [
                     [
                         'allow' => true,
-                        'roles' => ['@'],
-                        'matchCallback' => function () {
-                            return Yii::$app->user->identity->getIsAdmin();
-                        },
+                        'roles' => ['admin'],
                     ],
                 ],
             ],
@@ -109,12 +197,14 @@ class AdminController extends Controller
             'class'    => User::className(),
             'scenario' => 'create',
         ]);
+        $event = $this->getUserEvent($user);
 
         $this->performAjaxValidation($user);
 
+        $this->trigger(self::EVENT_BEFORE_CREATE, $event);
         if ($user->load(Yii::$app->request->post()) && $user->create()) {
             Yii::$app->getSession()->setFlash('success', Yii::t('user', 'User has been created'));
-
+            $this->trigger(self::EVENT_AFTER_CREATE, $event);
             return $this->redirect(['update', 'id' => $user->id]);
         }
 
@@ -135,12 +225,14 @@ class AdminController extends Controller
         Url::remember('', 'actions-redirect');
         $user = $this->findModel($id);
         $user->scenario = 'update';
+        $event = $this->getUserEvent($user);
 
         $this->performAjaxValidation($user);
 
+        $this->trigger(self::EVENT_BEFORE_UPDATE, $event);
         if ($user->load(Yii::$app->request->post()) && $user->save()) {
             Yii::$app->getSession()->setFlash('success', Yii::t('user', 'Account details have been updated'));
-
+            $this->trigger(self::EVENT_AFTER_UPDATE, $event);
             return $this->refresh();
         }
 
@@ -161,6 +253,7 @@ class AdminController extends Controller
         Url::remember('', 'actions-redirect');
         $user    = $this->findModel($id);
         $profile = $user->profile;
+        $event   = $this->getProfileEvent($profile);
 
         if ($profile == null) {
             $profile = Yii::createObject(Profile::className());
@@ -169,9 +262,11 @@ class AdminController extends Controller
 
         $this->performAjaxValidation($profile);
 
+        $this->trigger(self::EVENT_BEFORE_PROFILE_UPDATE, $event);
+
         if ($profile->load(Yii::$app->request->post()) && $profile->save()) {
             Yii::$app->getSession()->setFlash('success', Yii::t('user', 'Profile details have been updated'));
-
+            $this->trigger(self::EVENT_AFTER_PROFILE_UPDATE, $event);
             return $this->refresh();
         }
 
@@ -229,7 +324,13 @@ class AdminController extends Controller
      */
     public function actionConfirm($id)
     {
-        $this->findModel($id)->confirm();
+        $model = $this->findModel($id);
+        $event = $this->getUserEvent($model);
+
+        $this->trigger(self::EVENT_BEFORE_CONFIRM, $event);
+        $model->confirm();
+        $this->trigger(self::EVENT_AFTER_CONFIRM, $event);
+
         Yii::$app->getSession()->setFlash('success', Yii::t('user', 'User has been confirmed'));
 
         return $this->redirect(Url::previous('actions-redirect'));
@@ -248,7 +349,11 @@ class AdminController extends Controller
         if ($id == Yii::$app->user->getId()) {
             Yii::$app->getSession()->setFlash('danger', Yii::t('user', 'You can not remove your own account'));
         } else {
-            $this->findModel($id)->delete();
+            $model = $this->findModel($id);
+            $event = $this->getUserEvent($model);
+            $this->trigger(self::EVENT_BEFORE_DELETE, $event);
+            $model->delete();
+            $this->trigger(self::EVENT_AFTER_DELETE, $event);
             Yii::$app->getSession()->setFlash('success', Yii::t('user', 'User has been deleted'));
         }
 
@@ -267,12 +372,17 @@ class AdminController extends Controller
         if ($id == Yii::$app->user->getId()) {
             Yii::$app->getSession()->setFlash('danger', Yii::t('user', 'You can not block your own account'));
         } else {
-            $user = $this->findModel($id);
+            $user  = $this->findModel($id);
+            $event = $this->getUserEvent($user);
             if ($user->getIsBlocked()) {
+                $this->trigger(self::EVENT_BEFORE_UNBLOCK, $event);
                 $user->unblock();
+                $this->trigger(self::EVENT_AFTER_UNBLOCK, $event);
                 Yii::$app->getSession()->setFlash('success', Yii::t('user', 'User has been unblocked'));
             } else {
+                $this->trigger(self::EVENT_BEFORE_BLOCK, $event);
                 $user->block();
+                $this->trigger(self::EVENT_AFTER_BLOCK, $event);
                 Yii::$app->getSession()->setFlash('success', Yii::t('user', 'User has been blocked'));
             }
         }