+++ controllers/RegistrationController.php
@@ -11,11 +11,15 @@
 
 namespace dektrium\user\controllers;
 
+use dektrium\user\events\ConnectEvent;
+use dektrium\user\events\FormEvent;
+use dektrium\user\events\UserEvent;
 use dektrium\user\Finder;
 use dektrium\user\models\RegistrationForm;
 use dektrium\user\models\ResendForm;
 use dektrium\user\models\User;
 use dektrium\user\traits\AjaxValidationTrait;
+use dektrium\user\traits\EventTrait;
 use Yii;
 use yii\filters\AccessControl;
 use yii\web\Controller;
@@ -32,6 +36,55 @@ use yii\web\NotFoundHttpException;
 class RegistrationController extends Controller
 {
     use AjaxValidationTrait;
+    use EventTrait;
+
+    /**
+     * Event is triggered after creating RegistrationForm class.
+     * Triggered with \dektrium\user\events\FormEvent.
+     */
+    const EVENT_BEFORE_REGISTER = 'beforeRegister';
+
+    /**
+     * Event is triggered after successful registration.
+     * Triggered with \dektrium\user\events\FormEvent.
+     */
+    const EVENT_AFTER_REGISTER = 'afterRegister';
+
+    /**
+     * Event is triggered before connecting user to social account.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_CONNECT = 'beforeConnect';
+
+    /**
+     * Event is triggered after connecting user to social account.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_CONNECT = 'afterConnect';
+
+    /**
+     * Event is triggered before confirming user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_BEFORE_CONFIRM = 'beforeConfirm';
+
+    /**
+     * Event is triggered before confirming user.
+     * Triggered with \dektrium\user\events\UserEvent.
+     */
+    const EVENT_AFTER_CONFIRM = 'afterConfirm';
+
+    /**
+     * Event is triggered after creating ResendForm class.
+     * Triggered with \dektrium\user\events\FormEvent.
+     */
+    const EVENT_BEFORE_RESEND = 'beforeResend';
+
+    /**
+     * Event is triggered after successful resending of confirmation email.
+     * Triggered with \dektrium\user\events\FormEvent.
+     */
+    const EVENT_AFTER_RESEND = 'afterResend';
 
     /** @var Finder */
     protected $finder;
@@ -77,10 +130,16 @@ class RegistrationController extends Controller
 
         /** @var RegistrationForm $model */
         $model = Yii::createObject(RegistrationForm::className());
+        $event = $this->getFormEvent($model);
+
+        $this->trigger(self::EVENT_BEFORE_REGISTER, $event);
 
         $this->performAjaxValidation($model);
 
         if ($model->load(Yii::$app->request->post()) && $model->register()) {
+
+            $this->trigger(self::EVENT_AFTER_REGISTER, $event);
+
             return $this->render('/message', [
                 'title'  => Yii::t('user', 'Your account has been created'),
                 'module' => $this->module,
@@ -117,8 +176,13 @@ class RegistrationController extends Controller
             'email'    => $account->email,
         ]);
 
+        $event = $this->getConnectEvent($account, $user);
+
+        $this->trigger(self::EVENT_BEFORE_CONNECT, $event);
+
         if ($user->load(Yii::$app->request->post()) && $user->create()) {
             $account->connect($user);
+            $this->trigger(self::EVENT_AFTER_CONNECT, $event);
             Yii::$app->user->login($user, $this->module->rememberFor);
             return $this->goBack();
         }
@@ -147,7 +211,13 @@ class RegistrationController extends Controller
             throw new NotFoundHttpException();
         }
 
+        $event = $this->getUserEvent($user);
+
+        $this->trigger(self::EVENT_BEFORE_CONFIRM, $event);
+
         $user->attemptConfirmation($code);
+
+        $this->trigger(self::EVENT_AFTER_CONFIRM, $event);
 
         return $this->render('/message', [
             'title'  => Yii::t('user', 'Account confirmation'),
@@ -169,10 +239,16 @@ class RegistrationController extends Controller
 
         /** @var ResendForm $model */
         $model = Yii::createObject(ResendForm::className());
+        $event = $this->getFormEvent($model);
+
+        $this->trigger(self::EVENT_BEFORE_RESEND, $event);
 
         $this->performAjaxValidation($model);
 
         if ($model->load(Yii::$app->request->post()) && $model->resend()) {
+
+            $this->trigger(self::EVENT_AFTER_RESEND, $event);
+
             return $this->render('/message', [
                 'title'  => Yii::t('user', 'A new confirmation link has been sent'),
                 'module' => $this->module,