+++ migrations/Migration.php
@@ -29,9 +29,11 @@ class Migration extends \yii\db\Migration
     public function init()
     {
         parent::init();
-		
+
         switch (Yii::$app->db->driverName) {
             case 'mysql':
+                $this->tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
+                break;
             case 'pgsql':
                 $this->tableOptions = null;
                 break;