+++ migrations/m150623_212711_fix_username_notnull.php
@@ -16,7 +16,11 @@ class m150623_212711_fix_username_notnull extends Migration
 {
     public function up()
     {
-        $this->alterColumn('{{%user}}', 'username', Schema::TYPE_STRING . '(255) NOT NULL');
+        if ($this->db->driverName === 'pgsql') {
+            $this->alterColumn('{{%user}}', 'username', 'SET NOT NULL');
+        } else {
+            $this->alterColumn('{{%user}}', 'username', Schema::TYPE_STRING . '(255) NOT NULL');
+        }
     }
 
     public function down()