+++ models/RecoveryForm.php
@@ -13,8 +13,10 @@ namespace dektrium\user\models;
 
 use dektrium\user\Finder;
 use dektrium\user\Mailer;
+use dektrium\user\traits\ModuleTrait;
 use Yii;
 use yii\base\Model;
+use yii\helpers\ArrayHelper;
 
 /**
  * Model for collecting data on password recovery.
@@ -25,6 +27,7 @@ use yii\base\Model;
  */
 class RecoveryForm extends Model
 {
+    use ModuleTrait;
     /** @var string */
     public $email;
 
@@ -33,9 +36,6 @@ class RecoveryForm extends Model
 
     /** @var User */
     protected $user;
-
-    /** @var \dektrium\user\Module */
-    protected $module;
 
     /** @var Mailer */
     protected $mailer;
@@ -50,7 +50,6 @@ class RecoveryForm extends Model
      */
     public function __construct(Mailer $mailer, Finder $finder, $config = [])
     {
-        $this->module = Yii::$app->getModule('user');
         $this->mailer = $mailer;
         $this->finder = $finder;
         parent::__construct($config);
@@ -68,10 +67,11 @@ class RecoveryForm extends Model
     /** @inheritdoc */
     public function scenarios()
     {
-        return [
+        $scenarios = parent::scenarios();
+        return ArrayHelper::merge($scenarios, [
             'request' => ['email'],
             'reset'   => ['password'],
-        ];
+        ]);
     }
 
     /** @inheritdoc */