+++ models/SettingsForm.php
@@ -14,6 +14,7 @@ namespace dektrium\user\models;
 use dektrium\user\helpers\Password;
 use dektrium\user\Mailer;
 use dektrium\user\Module;
+use dektrium\user\traits\ModuleTrait;
 use Yii;
 use yii\base\Model;
 
@@ -26,6 +27,8 @@ use yii\base\Model;
  */
 class SettingsForm extends Model
 {
+    use ModuleTrait;
+
     /** @var string */
     public $email;
 
@@ -37,9 +40,6 @@ class SettingsForm extends Model
 
     /** @var string */
     public $current_password;
-
-    /** @var Module */
-    protected $module;
 
     /** @var Mailer */
     protected $mailer;
@@ -61,7 +61,6 @@ class SettingsForm extends Model
     public function __construct(Mailer $mailer, $config = [])
     {
         $this->mailer = $mailer;
-        $this->module = Yii::$app->getModule('user');
         $this->setAttributes([
             'username' => $this->user->username,
             'email'    => $this->user->unconfirmed_email ?: $this->user->email,