+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.0.0-alpha [Work in progress]
+
+- `enh` Support for custom `admin` role via access control rule out of the box #510 (dmeroff)
+- `fix` Removed ability to register a new account via social network when registration is disabled #512 (dmeroff)
+- `fix` Added check if profile model exists in Settings controller #497 (dmeroff)
+- `enh` Added event triggers to most of actions #411 (dmeroff)
+- `enh` Added LinkedIn client #496 (SamMousa)
+
 ## 0.9.5 [27 September 2015]
 
 - `fix` Don't set mailer subject in bootstrap #451 (dmeroff)